+++ app/models/config.py
@@ -306,6 +306,7 @@ class Config:
         self.dark = False
         self.nojs = False
         self.near = ''
+        self.alts = False
         self.new_tab = False
         self.get_only = False
